revise eulerian cases
