+++ tests/user_examples/test_2d_FVM_double_mach_reflection/2d_FVM_double_mach_reflection.cpp
@@ -1,6 +1,6 @@
 /**
  * @file 	2d_FVM_double_mach_reflection.cpp
- * @brief 	This is the compressible test for the realizaiton of FVM in the SPHinXsys.
+ * @brief 	This is the compressible test for the realization of FVM in the SPHinXsys.
  * @details We consider a double mach reflection case.
  * @author 	Zhentong Wang and Xiangyu Hu
  */
@@ -12,7 +12,7 @@ using namespace SPH;
 //----------------------------------------------------------------------
 int main(int ac, char *av[])
 {
-    // read data from ANASYS mesh.file
+    // read data from ANSYS mesh.file
     readMeshFile read_mesh_data(double_mach_reflection_mesh_fullpath);
     //----------------------------------------------------------------------
     //	Build up the environment of a SPHSystem.
@@ -51,8 +51,9 @@ int main(int ac, char *av[])
     the value is larger, the numerical dissipation larger*/
     InteractionWithUpdate<Integration1stHalfHLLCRiemann> pressure_relaxation(water_block_inner);
     InteractionWithUpdate<Integration2ndHalfHLLCRiemann> density_relaxation(water_block_inner);
-    // Visuallization in FVM with date in cell.
-    BodyStatesRecordingInMeshToVtp write_real_body_states(io_environment, sph_system.real_bodies_, read_mesh_data.elements_nodes_connection_, read_mesh_data.point_coordinates_2D_);
+    // Visualization in FVM with date in cell.
+    BodyStatesRecordingInMeshToVtp write_real_body_states(
+        io_environment, sph_system.real_bodies_, read_mesh_data.elements_nodes_connection_, read_mesh_data.point_coordinates_2D_);
     //----------------------------------------------------------------------
     //	Prepare the simulation with case specified initial condition if necessary.
     //----------------------------------------------------------------------
@@ -107,7 +108,6 @@ int main(int ac, char *av[])
     TimeInterval tt;
     tt = t4 - t1 - interval;
     cout << "Total wall time for computation: " << tt.seconds() << " seconds." << endl;
-
 
     return 0;
 }
