+++ tests/user_examples/test_2d_FVM_double_mach_reflection/CMakeLists.txt
@@ -16,11 +16,7 @@ target_sources(${PROJECT_NAME} PRIVATE ${DIR_SRCS})
 target_link_libraries(${PROJECT_NAME} extra_sources_2d)
 set_target_properties(${PROJECT_NAME} PROPERTIES VS_DEBUGGER_WORKING_DIRECTORY "${EXECUTABLE_OUTPUT_PATH}")
 
-add_test(NAME ${PROJECT_NAME}_particle_relaxation
-        COMMAND ${PROJECT_NAME} --relax=true --state_recording=${TEST_STATE_RECORDING}
-        WORKING_DIRECTORY ${EXECUTABLE_OUTPUT_PATH})
 add_test(NAME ${PROJECT_NAME}
-        COMMAND ${PROJECT_NAME}  --reload=true --state_recording=${TEST_STATE_RECORDING}
+        COMMAND ${PROJECT_NAME} --state_recording=${TEST_STATE_RECORDING}
         WORKING_DIRECTORY ${EXECUTABLE_OUTPUT_PATH})
-set_tests_properties(${PROJECT_NAME} PROPERTIES DEPENDS "${PROJECT_NAME}_particle_relaxation")
-set_tests_properties(${PROJECT_NAME} PROPERTIES LABELS "periodic boundary, Eulerian")
+set_tests_properties(${PROJECT_NAME} PROPERTIES LABELS "compressible flow, Eulerian")
